gem-private-server-py/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
botasaurus==3.2.25
botasaurus_proxy_authentication==1.0.12
cachetools==5.3.3
capsolver_extension_python==1.0.1
casefy==0.1.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chrome_extension_python==1.0.1
chromedriver-autoinstaller-fix==1.0.5
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
dataclasses-json==0.6.4
dateutils==0.6.12
fastapi==0.110.0
frozenlist==1.4.1
geonamescache==2.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
javascript_fixes==1.1.18
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-google-genai==0.0.9
langsmith==0.1.10
lxml==5.1.0
marshmallow==3.21.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.2.3
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
ujson==5.9.0
Unidecode==1.3.8
urllib3==1.26.18
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
youtube-search-python==1.6.6